# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UrlInput {
  /** @minLength 1 */
  url: string;
}

export interface BatchUrlInput {
  /**
     * @minItems 1
     * @maxItems 20
     */
  urls: string[];
}

export type AnalysisResultDetails = { [key: string]: unknown };

export interface AnalysisResult {
  url: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  verdict: string;
  riskLevel: string;
  flags: string[];
  details?: AnalysisResultDetails;
}

export type ScanRecordDetails = { [key: string]: unknown };

export interface ScanRecord {
  id: number;
  url: string;
  score: number;
  verdict: string;
  riskLevel: string;
  flags: string[];
  details?: ScanRecordDetails;
  createdAt: string;
}

export interface HistoryResponse {
  scans: ScanRecord[];
  total: number;
}

export interface HistoryStats {
  total: number;
  safe: number;
  suspicious: number;
  likelyPhishing: number;
  phishing: number;
  avgScore: number;
}

export interface DeleteResult {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type GetHistoryParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
riskLevel?: string;
};