# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: phishing
    description: Phishing URL detection
  - name: history
    description: Scan history
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /analyze:
    post:
      operationId: analyzeUrl
      tags: [phishing]
      summary: Analyze a URL for phishing indicators
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UrlInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyze/batch:
    post:
      operationId: analyzeUrlBatch
      tags: [phishing]
      summary: Analyze multiple URLs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchUrlInput"
      responses:
        "200":
          description: List of analysis results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /history:
    get:
      operationId: getHistory
      tags: [history]
      summary: Get recent scan history
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - in: query
          name: riskLevel
          schema:
            type: string
      responses:
        "200":
          description: List of past scans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
  /history/{id}:
    delete:
      operationId: deleteScan
      tags: [history]
      summary: Delete a scan record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /history/stats:
    get:
      operationId: getHistoryStats
      tags: [history]
      summary: Aggregate stats across all scans
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryStats"
components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    UrlInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
          minLength: 1
    BatchUrlInput:
      type: object
      required: [urls]
      properties:
        urls:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 20
    AnalysisResult:
      type: object
      required: [url, score, verdict, riskLevel, flags]
      properties:
        url:
          type: string
        score:
          type: integer
          minimum: 0
          maximum: 100
        verdict:
          type: string
        riskLevel:
          type: string
        flags:
          type: array
          items:
            type: string
        details:
          type: object
          additionalProperties: true
    ScanRecord:
      type: object
      required: [id, url, score, verdict, riskLevel, flags, createdAt]
      properties:
        id:
          type: integer
        url:
          type: string
        score:
          type: integer
        verdict:
          type: string
        riskLevel:
          type: string
        flags:
          type: array
          items:
            type: string
        details:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    HistoryResponse:
      type: object
      required: [scans, total]
      properties:
        scans:
          type: array
          items:
            $ref: "#/components/schemas/ScanRecord"
        total:
          type: integer
    HistoryStats:
      type: object
      required: [total, safe, suspicious, likelyPhishing, phishing, avgScore]
      properties:
        total:
          type: integer
        safe:
          type: integer
        suspicious:
          type: integer
        likelyPhishing:
          type: integer
        phishing:
          type: integer
        avgScore:
          type: number
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string