# lib/api-zod/src/generated/types/historyResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanRecord } from './scanRecord';

export interface HistoryResponse {
  scans: ScanRecord[];
  total: number;
}