# lib/api-zod/src/generated/types/historyStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryStats {
  total: number;
  safe: number;
  suspicious: number;
  likelyPhishing: number;
  phishing: number;
  avgScore: number;
}