# lib/api-zod/src/generated/types/scanRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanRecordDetails } from './scanRecordDetails';

export interface ScanRecord {
  id: number;
  url: string;
  score: number;
  verdict: string;
  riskLevel: string;
  flags: string[];
  details?: ScanRecordDetails;
  createdAt: Date;
}